dplayx: Implemented wide version of EnumConnections
  Now both ANSI and wide versions of EnumConnections call an auxiliary
function with parameter bAnsi=(TRUE|FALSE).
  This approach is used instead of the traditional Wine approach because
in the middle of the function we need to call a callback function with
either ANSI or wide strings.
  The code of the auxiliar function is basically the code of the
previous ANSI version cleaned up, as the previous code implemented
Lobby Providers support with a copy/paste and a few changes.
  This patch fixes connection enumeration in GTA2.

  Changed DP_BuildSPCompoundAddr to be more generic, and renamed
it to DP_BuildCompoundAddr.

Based on patch by Ismael Barros <razielmine@gmail.com>